Use resolve helper.
In the case where there is no Promise that needs to be resolved, this
resolve helper returns a `then`able object that calls the function
immediately. This means that the behaviour remains backwards compatible
(i.e. no deffered behaviour) for developers who choose not to use
Promises.

It may be worth exposing this helper as someone may want to use await.
I.e.

    Backbone.resolve = function(promise) {
      return {
        then: function(fn) {
	  return await Promise.resolve(promise).then(fn);
	}
      };
    }